Added support for text inputs countUps
Added a function "formatPrint" to the countUp object that checks if the
"this.d" object is a an input. If it is, it applies the countUp frame to
the "this.d.val" instead of the "this.d.innerHTML." If it is not an
input, the function applies the frame to the innerHTML just as before.

Replaced all "this.d.innerHTML =
this.formatNumber(this.frameVal.toFixed(this.decimals));" calls with
this function.